reformat code and minor fixes
remove mal-coded matchPartialKey, replace it with getMatchLength

finish serializer basically

finish ser and deser

a bug wasted one night: in SEARTNode#getChildrenWithKeys Node48 should return all items where keys[pos] >= 1, if the equal symbol is dropped, there could be many issues caused by an incomplete Node48. It would be hard to debug because Node48 is always mingled with a bunch of keys (>16) which makes it difficult to observe the error.

The root reason is lack of envelop. If Node48 and its brothers are always handled with member methods, it would be more rare to get this problem.

remove redundant getKeyedChildren, better insert, unsign bytes in Node4/16